discord-bot/requirements.txt
discord.py==2.3.2
google-cloud-storage==2.18.2
aiohttp==3.10.5